+++ cmd/tools/pmanager/main.go
@@ -28,7 +28,7 @@ func main() {
 		login   = []byte(os.Args[2])
 	)
 
-	keyBuilder := keybuilder.NewKeyBuilder(workSize, login)
+	keyBuilder := keybuilder.NewKeyBuilder(1<<workSize, login)
 	extendedKey := keyBuilder.Build(readUntilEOF("> "))
 
 	passBytes := hashing.NewHMACSHA256Hasher(extendedKey, service).ToBytes()