+++ pkg/network/message/message_test.go
@@ -20,7 +20,7 @@ const (
 	tcHead       = 12345
 	tcBody       = "hello, world!"
 	tcNetworkKey = "network_key"
-	tcProof      = 875
+	tcProof      = 1849
 )
 
 type sInvalidPayload struct{}
@@ -89,7 +89,7 @@ func TestMessage(t *testing.T) {
 
 	msg3 := NewMessage(sett, pld).(*sMessage)
 	msg3.fHash = random.NewStdPRNG().GetBytes(hashing.CSHA256Size)
-	msg3.fProof = puzzle.NewPoWPuzzle(testutils.TCWorkSize, cPuzzleIterN).ProofBytes(msg3.fHash)
+	msg3.fProof = puzzle.NewPoWPuzzle(testutils.TCWorkSize, 1).ProofBytes(msg3.fHash)
 	if _, err := LoadMessage(sett, msg3.ToBytes()); err == nil {
 		t.Error("success load with invalid hash")
 		return